# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Fitness Coach API Server
  description: OpenAPI 3.0 specification for FastAPI server and WebSocket inference endpoints in AI Fitness Coach.
  version: 2.0.0
servers:
  - url: http://127.0.0.1:8000
    description: Local Development Server
paths:
...
  /api/exercises/{name}:
    get:
      summary: Get exercise metadata
      description: Returns detailed configuration, alignment settings, thresholds, and coaching rules for a specific exercise.
      operationId: getExerciseInfo
      parameters:
        - name: name
          in: path
          required: true
          description: Name of the target exercise (e.g. Bicep Curl, Squats)
          schema:
            type: string
      responses:
        '200':
          description: Exercise metadata returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExerciseConfig'
        '404':
          description: Exercise not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
...
components:
  schemas:
...
    ExerciseConfig:
      type: object
      properties:
        type:
          type: string
          example: REP_BASED
        category:
          type: string
          example: UPPER BODY
        primary_joint:
          type: string
          example: elbow
        keypoints:
          type: array
          items:
            type: string
          example: ["left_shoulder", "left_elbow", "left_wrist"]
        alt_keypoints:
          type: array
          items:
            type: string
          example: ["right_shoulder", "right_elbow", "right_wrist"]
        down_threshold:
          type: number
          format: float
          example: 150.0
        up_threshold:
          type: number
          format: float
          example: 90.0
        target_reps:
          type: integer
          example: 10
        cooldown:
          type: number
          format: float
          example: 0.8
        alignment:
          type: object
        checks:
          type: object
...
    HTTPError:
      type: object
      properties:
        detail:
          type: string
          example: Exercise not found